Moved MovieClickListener to MainActivity.kt
