+++ app/src/main/java/academy/android/mymovie/adapter/MovieAdapter.kt
@@ -80,4 +80,5 @@ class MovieViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
 
 interface MovieClickInterface {
     fun onMovieClick(position: Int)
+    fun onBackClick()
 }
